Add Dockerfile to build rootfs image from scratch
This Dockerfile uses an Alpine base image to set up
the initial file system by copying the project files
into a / directory. It cleans up unnecessary files like
.git, .DS_Store, .github, .gitignore, and Dockerfile
itself to ensure they are not included in the build.
The file ownership is changed to root, and then the
prepared filesystem is copied from the Alpine builder
to a scratch-based final image. This setup provides a c
lean and minimal rootfs image tailored for production use.